define SparqleResponse in types/
