+++ web/pages/rdfs/_class/_key.vue
@@ -25,44 +25,7 @@
 import Vue from 'vue'
 import SparqlResponseTable from '~/components/SparqlResponseTable.vue'
 
-interface RdfTerms1 {
-  type: 'uri'
-  value: string
-}
-
-interface RdfTerms2 {
-  type: 'literal'
-  value: string
-}
-
-interface RdfTerms3 {
-  type: 'literal'
-  value: string
-  'xml:lang': string
-}
-
-interface RdfTerms4 {
-  type: 'literal'
-  value: string
-  datatype: string
-}
-
-interface RdfTerms5 {
-  type: 'bnode'
-  value: string
-}
-
-interface SparqleResponse {
-  head: {
-    vars: string[]
-    link: string[]
-  }
-  results: {
-    bindings: {
-      [key: string]: RdfTerms1 | RdfTerms2 | RdfTerms3 | RdfTerms4 | RdfTerms5
-    }[]
-  }
-}
+import { SparqleResponse } from '~/types/SparqleResponse.d.ts'
 
 export default Vue.extend({
   components: { SparqlResponseTable },
